finish a one-batch test
